requirements.txt
fastapi
uvicorn[standard]
tensorflow-cpu==2.12.0
numpy==1.23.5
pillow
python-multipart
